requirements.txt
sseclient_py==1.7.2